requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==1.12.0
pymupdf==1.23.26
langchain==0.1.9
langchain-openai==0.0.8
faiss-cpu==1.7.4
python-dotenv==1.0.1
nltk==3.8.1
spacy==3.7.4
pandas==2.2.1
numpy==1.26.4
gunicorn==21.2.0 